pyproject.toml
[project]
name = "webuitestautomation"
version = "0.1.0"
description = ""
authors = [
    {name = "software groups",email = "admin@sfgroups.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright (>=1.52.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "webuitestautomation", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-playwright = "^0.7.0"
pytest-html = "^4.1.1"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
